requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
plotly>=5.13.0
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
